requirements.txt
python-dotenv==1.0.1
click==8.1.7
ensure==1.0.4
numpy==2.1.3
pandas==2.2.3
setuptools==75.5.0
scikit-learn==1.5.2
mlflow==2.17.2
hydra-core==1.3.2
typer==0.13.0
dvc[s3]==3.56.0
fastapi==0.115.6
# airflow
pytest==8.3.3
dnspython==2.7.0
flake8==7.1.1
mypy==1.13.0
pytest-cov==6.0.0
pre-commit==4.0.1
isort==5.13.2